Fix 'Use of 32-bit time_t' Coverity issues.

+++ src/examples/encrypt.c
@@ -82,7 +82,7 @@ ffi_encrypt()
     /* setup encryption parameters */
     rnp_op_encrypt_set_armor(encrypt, true);
     rnp_op_encrypt_set_file_name(encrypt, "message.txt");
-    rnp_op_encrypt_set_file_mtime(encrypt, time(NULL));
+    rnp_op_encrypt_set_file_mtime(encrypt, (uint32_t) time(NULL));
     rnp_op_encrypt_set_compression(encrypt, "ZIP", 6);
     rnp_op_encrypt_set_cipher(encrypt, RNP_ALGNAME_AES_256);
     rnp_op_encrypt_set_aead(encrypt, "None");
